Move util files into a subfolder
