+++ bril-cpp/src/util/casting.hpp
