+++ bril-cpp/src/util/string_pool.hpp
